Rewrites 'graph' module to do less work.
The 'tsort' function now requires an explicit parameter to obtain adjacencies
rather than invoking a 'dependencies()' method, and offers control over how the
adjacencies should be normalized before traversal.  The 'tsort' function also
reports the nodes in the cycle if such a cycle exists.  Finally the confusion
between 'tsort' and 'traverse' is eliminated.