bdemeta: add support for PkgConfig packages
Introduce support for PkgConfig packages.  A target name found in the
"pkg_configs" section of the configuration file will create a CMake
target whose include directories, dependencies and compiler options are
sourced from the associated PkgConfig package.

In order to support testing this code path, the CMake test driver is
significantly refactored.  In particular, a basic parser for CMake "if"
commands is introduced, and tests are added for the parser.