Separate roots into bde_roots and cmake_dirs
The configuration file is now required to explicitly delineate which
targets may be searched in BDE repository roots against targets that are
CMake projects.

The primary motivation for this is to allow CMake projects to have
arbitrary directory names.  Previously only CMake projects that were
containined in a directory matching the target name (or via a
'thirdparty' subdirectory) were supported.