bdemeta: paths relative to configuration file
Relative paths in the 'roots' configuration option are now treated
relative to the path of the configuration file, and not relative to the
current working directory.

This makes it more convenient to leave configuration files in source
repositories, and to invoke 'bdemeta' from some arbitrary build
directory without needing to amend the configuration file.