bdemeta/__main__: use argparse
Use argparse instead of custom argument parsing. argparse has not been
needed thus far as no options are present in the CLI of 'bdemeta'. This
change anticipates the addition of options.

Drop support for Python 3.6 since this uses a feature of the 'argparse'
module which is only present in Python 3.7 and newer.