cmake_parser: search for longest match
When searching for arguments specified to command, ensure the generated
CMake list contains at least as many arguments as specified in the
search.

Implementation note: relying on Python's built-in 'zip' function is
insufficient as it searches for the shortest match.  Notably,
'itertools.zip_longest' is also not appropriate since it searches for
the longest match and the generated CMake lists may contain more
arguments than we are interested in searching for.