bdemeta: make plugin tests a global config
Experimentation with the per-target 'plugin_tests' configuration value
has shown that it is instead preferable to have all tests built as
shared libraries instead of a subset.

This removes the per-target configurability of plugin tests, and instead
supports a '-p' flag for the 'cmake' subcommand which builds all tests
as plugins instead.

NOTE: this is a breaking change, since support for the per-target
configurability is removed.